Fixed an issue with timestamps being truncated to 32bit, this was creating completely different numbers making the comparison between current time and last update time whack.

+++ src/LoginServer/LoginManager.h
@@ -114,8 +114,8 @@ private:
     ServerDataList          mServerDataList;
     bool                    mSendServerList;
 
-    uint32                  mLastStatusQuery;
-    uint32                  mLastHeartbeat;
+    uint64                  mLastStatusQuery;
+    uint64                  mLastHeartbeat;
     uint32					mNumClientsProcessed;
 
     boost::pool<boost::default_user_allocator_malloc_free>	mLoginClientPool;
